% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Example Bank Data to Use in Multiserve simulation}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{arrival_time}{arrival time of customer}
\item{service_time}{length of service for customer}
}
}
\usage{
bank
}
\description{
Example Bank Data to Use in Multiserve simulation
}
\examples{
data(bank)
str(bank)
}
\keyword{datasets}
