% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Queuing System Simulation}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Arrival time column of dataset}

\item{ServiceTimes}{Service time of dataset}

\item{NumServers}{Specify the number of servers for the simulation}
}
\value{
Tibble with 4 variables; Arrivals, ServiceBegins, ChosenServer, ServiceEnds
}
\description{
Given the arrival & service time columns of a dataset and the number
of servers, this function simulates when the service begins and ends along with
the specific server during the period.
}
\examples{
data(bank)
str(bank)
Multiserver(bank$arrival_time, bank$service_time, NumServers = 3)
}
